refactor: 불필요한 User 인터페이스 삭제 후 AbstractUser 추상클래스의 이름을 User로 변경
