+++ src/main/java/blackjack/domain/user/Player.java
@@ -1,6 +1,6 @@
 package blackjack.domain.user;
 
-public class Player extends AbstractUser {
+public class Player extends User {
     private Player(String name) {
         super(name);
     }
