+++ src/test/java/blackjack/domain/user/UserTest.java
@@ -11,7 +11,7 @@ import java.util.Arrays;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class AbstractUserTest {
+public class UserTest {
     User user;
 
     @BeforeEach
